feat: Add proposals page
I already added the following:
- Factories
- Seeders

Additionally I added a feature where the client can accept or reject a
proposal. The problem I have right now is that when I accept or reject a
proposal, it reloads the entire page. I don't how to solve that one yet.
But hey, at least it works.